api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
twilio==9.3.3
bcrypt==4.2.0
PyJWT==2.9.0
pydantic==2.9.2
python-multipart==0.0.9
